Allocate empty object when doing ethereum login
When adding an ethereum coin, 0 balance is shown in a wallet, and an error message is shown in the console. The error message is that is tried to add a property to api.eth.connect when it contains a null value. For avoiding this situation, api.eth.connect must be assigned to an empty object when starting to work with a Ethereum wallet, this is, at login.